requirements.txt
werkzeug==0.14.1
tensorflow
flask
cassandra-driver
pillow